+++ packages/macros/src/dependency-satisfies.ts
